% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_xs_module.R
\name{draw_xs_server}
\alias{draw_xs_server}
\title{Draw XS Server Function}
\usage{
draw_xs_server(id)
}
\arguments{
\item{id}{}
}
\value{
cross section reactive
}
\description{
Draw XS Server Function
}
