% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_web_crs.R
\name{fix_web_crs}
\alias{fix_web_crs}
\title{Fix Web CRS}
\usage{
fix_web_crs(obj)
}
\arguments{
\item{obj}{sf or terra object; The geospatial object to be corrected.}
}
\value{
the geospatial object with CRS correctly defined to match feature 
coordinate geometry.
}
\description{
Fixes mis-specified coordinate reference system (CRS) with 
             respect to the geospatial object's coordinate units. Ensures
             that the CRS matches the feature geometry units.
}
\details{
Geospatial features generated using the GeoJSON format often 
contain a CRS that does not match the feature geometry coordinate units. 
The GeoJSON format made the choice to not even try: 
[RFC 7946 - The GeoJSON Format](https://datatracker.ietf.org/doc/html/rfc7946#page-12).
}
