% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terrain_map.R
\name{get_terrain_map}
\alias{get_terrain_map}
\title{Get a Terrain Map}
\usage{
get_terrain_map(xs, dem)
}
\arguments{
\item{xs}{sf object, The cross sections}

\item{dem}{SpatRaster object, The digital elevation model.}
}
\value{
a tmap object
}
\description{
Get a terrain map as a tmap object.
}
