% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terrain_tmap.R
\name{get_terrain_tmap}
\alias{get_terrain_tmap}
\title{Get a Terrain tmap}
\usage{
get_terrain_tmap(xs, dem)
}
\arguments{
\item{xs}{sf object, The cross sections}

\item{dem}{SpatRaster object, The digital elevation model.}
}
\value{
a tmap object
}
\description{
Get a terrain map as a tmap object.
}
