% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_test_data.R
\name{save_test_data}
\alias{save_test_data}
\title{Save Test Data}
\usage{
save_test_data(data, filename)
}
\arguments{
\item{data}{sf or terra SpatRaster object; data to be saved in the 
package for testing}

\item{filename}{filename}
}
\value{
nothing, saves data to test folder
}
\description{
Function for saving spatial objects created within a Shiny app 
             to support package testing.
}
