% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_dimensions_table.R
\name{xs_dimensions_table}
\alias{xs_dimensions_table}
\title{Create a Cross Section Dimensions Table}
\usage{
xs_dimensions_table(xs_pts, xs_number, bf_estimate, regions)
}
\arguments{
\item{xs_pts}{sf; A cross section lines feature class.}

\item{xs_number}{integer; The cross section `Seq` number of the
requested cross section.}

\item{bf_estimate}{numeric; Detrended bankfull estimate (units:
detrended feet).}

\item{regions}{character vector; Regions to calculate hydraulic
dimensions for. See the `RegionalCurve` package for
a list of regions.}
}
\value{
a `gt` object
}
\description{
Creates a cross section dimensions table for the channel portion of the
specified cross section.
}
